deletador arrumado

+++ deletador.py
@@ -6,19 +6,20 @@ def deletar_aluno():
     try:
         lista() 
         aluno_id = int(input("\nID do aluno que deseja deletar: "))
-        resultado = executar("SELECT nome,  FROM aluno WHERE id = %s", (aluno_id,), fetch=True)
+        resultado = executar("SELECT nome FROM aluno WHERE id = %s", (aluno_id,), fetch=True)
         if not resultado:
             print("Erro: Aluno não encontrado.")
             return
-        executar(
-        "DELETE FROM aluno WHERE id = %s",
-        (aluno_id,)
-        )
-        print("Aluno removido com sucesso.")
+        if resultado:
+            executar(
+            "DELETE FROM aluno WHERE id = %s",
+            (aluno_id,)
+            )
+            print("Aluno removido com sucesso.")
     except Exception as e:
-        print("Erro: aluno não achado")
+        print("Erro:", e)
 def deletar_nota():
-    """Permite ao usuário escolher uma nota específica para deletar (zerar) e recalcula as médias."""
+
     try:
         lista()
         aluno_id = int(input("\nDigite o ID do aluno que deseja deletar a nota: "))
